+++ client/src/app/dashboard/settings/company/BasicDetails.tsx
@@ -1,0 +1,11 @@
+import React from 'react'
+
+type Props = {}
+
+const BasicDetails = (props: Props) => {
+  return (
+    <div>BasicDetails</div>
+  )
+}
+
+export default BasicDetails
