+++ client/src/app/dashboard/settings/company/ManageInvites.tsx
@@ -1,0 +1,11 @@
+import React from 'react'
+
+type Props = {}
+
+const ManageInvites = (props: Props) => {
+  return (
+    <div>JoinRequests</div>
+  )
+}
+
+export default ManageInvites
