+++ client/src/app/dashboard/settings/company/ManageUsers.tsx
@@ -1,0 +1,11 @@
+import React from 'react'
+
+type Props = {}
+
+const ManageUsers = (props: Props) => {
+  return (
+    <div>ManageUsers</div>
+  )
+}
+
+export default ManageUsers
